Add a validator to check for valid domain names
You are only supposed to type the domain name, because the script tag
will autocomplete the other parts of the URL, such as the proxy.
Therefore, validate that the settings are using domains, otherwise
you could just type 'http://example.com/js' a proxy and it would set a
value for https://http://example.com/js/js/script.js, which would be a
mess.