Use default domain when the plugin starts
If you don't change the domain, it will use the one from the forum
installation path. This also prevents the plugin from being enabled
without a domain name.

Even if you don't change anything, the default tracking domain is going
to be the forum install, and the default proxy will be plausible.io.